@86epku02q #Created a custom dialog and reused it, and isolated mock data and validation.
