+++ apps/simulator/src/components/Dialog/ReportIssue/IssueMockData.ts
@@ -1,0 +1,8 @@
+import { SelectOption } from "@repo/core-library/components";
+
+export const IssueMockData: SelectOption[] = [
+  { label: 'Bug Problems', value: 'Bug Problems', xvalue: 1 },
+  { label: 'Simulator Not Working', value: 'Simulator Not Working', xvalue: 2 },
+  { label: 'Question does not load', value: 'Question does not load', xvalue: 3 },
+  { label: 'Button Not Working', value: 'Button Not Working', xvalue: 4 },
+];
